replace global resource with mysql config and update mysql module that use mysql config
