+++ include/aquarius/basic_server.hpp
@@ -5,8 +5,8 @@
 #include <aquarius/io_service_pool.hpp>
 #include <aquarius/logger.hpp>
 #include <aquarius/module/module_router.hpp>
+#include <aquarius/module/mysql_module.hpp>
 #include <aquarius/module/player_module.hpp>
-#include <aquarius/module/sql_module.hpp>
 
 using namespace std::chrono_literals;
