--- include/aquarius/resource/config_register.hpp
@@ -1,17 +1,0 @@
-#pragma once
-#include <aquarius/resource/global_resource.hpp>
-
-namespace aquarius
-{
-	template <typename T>
-	class config_register
-	{
-	public:
-		config_register()
-			: value_()
-		{}
-
-	private:
-		T value_;
-	};
-} // namespace aquarius
